add hostname
The code for acquiring the Windows target IP address is not accurate.  on my system, for example, it grabs an IPV6 address,  The `-first 1` in the command appears to be the issue.  I am adding a line to acquire the hostname as well.  We should try to write a more accurate line for obtaining the windows target IPV4 address.